typedef double REAL

+++ src/elements/DEM_ellip3d/ellip3d/realtypes.h
@@ -3,7 +3,7 @@
 
 namespace dem { 
 
-  typedef long double REAL;
+  typedef double REAL;
 
 }
